backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
Pillow==10.0.0
psycopg2-binary
python-dotenv==1.0.0
gunicorn==21.2.0
dj-database-url==2.1.0